Feature/orders 6 18 (#3)
* added docker; controller; model; tests

* implement customer and order controller; fix pets tests

* implement customer controller; fix pets tests

* merge; resolve merge conflicts

* add product controller - 19 06

---------

+++ src/main/java/com/pawilion/repository/PetRepository.java
@@ -3,4 +3,4 @@ package com.pawilion.repository;
 import com.pawilion.model.Pet;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
-public interface PetRepository extends ReactiveCrudRepository<Pet, Integer> {}
+public interface PetRepository extends ReactiveCrudRepository<Pet, Long> {}
